audit(2.9.7): u128 migration complete + G-01a/G-01g/H-02a gates
F-010: 18 src/test files updated to u128, burn.rs add_fees_burned u64 to u128
G-01g: test_reorg_6_blocks_orphan_scenario written and passing
H-02a: MAINNET_CONSTITUTION.md FROZEN SHA-256 c76aa00224a37ba52950f7ce9e2f72cfc87aa84f5599d3ea28e57538441a8d97
audit: 460 tests 0 failures, 95/102 items complete

+++ L1/core/src/mempool/pool.rs
@@ -266,7 +266,7 @@ mod tests {
     use super::*;
     use crate::tx::{Transaction, TxInput, TxOutput};
 
-    fn make_tx(id: &str, fee: u64, inputs: Vec<(&str, u32)>, outputs: Vec<u64>) -> Transaction {
+    fn make_tx(id: &str, fee: u64, inputs: Vec<(&str, u32)>, outputs: Vec<u128>) -> Transaction {
         Transaction {
             id: id.to_string(),
             version: 1,
